fix: renamed file paths
