+++ apps/webapp/src/script/components/cells/common/cellsStateInfo/cellsStateInfo.styles.ts
