+++ apps/webapp/src/script/components/cells/common/cellsStateInfo/cellsStateInfo.test.tsx
